Fixed not showing friend's posts
I realized that the association was not bidirectional.
The frienship model specified user_id and friend_id. To
Rails, it was a user to friend relationship, not friend to friend relationship.
Had to create another friendship model with user_id's switched upon accepting friend
request to simulate a bidirectional relationship.